Update solution for sigmoid-numpy

+++ sigmoid-numpy/sigmoid-numpy.py
@@ -4,6 +4,6 @@ def sigmoid(x):
     """
     Vectorized sigmoid function.
     """
-    x = np.array(x)
+    x = np.array(x, dtype=float)
     # Write code here
     return 1/(1+ np.exp(-x))
